move clang tidy fixes

+++ csrc/scheduler/mma_utils.cpp
@@ -861,7 +861,8 @@ void WarpMmaSwizzler::scheduleTMALoadForMma(
   // In the comments below I have kept K as the outer dimension. That is
   // just to have a concrete running example - it can be inner or outer.
 
-  int64_t num_ids_to_skip = tv->getLoopDomain().size() - 2;
+  int64_t num_ids_to_skip =
+      static_cast<int64_t>(tv->getLoopDomain().size() - 2);
   NVF_ERROR(num_ids_to_skip >= 0);
   if (swizzle == MmaInputSmemSwizzle::None) {
     // For no-swizzle case, the entire tile are divided into 8x8 core matrices,
